--- src/components/AddArticleItem/styles.scss
@@ -1,40 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.create-article {
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  width: 100%;
-  max-width: 115.2rem;
-  margin: 0 auto;
-
-  &__header {
-    display: flex;
-    align-items: center;
-    margin-top: 2rem;
-
-    border-bottom: $border-btn-primary;
-    & > * {
-      margin-bottom: 2rem;
-    }
-    h1 {
-      margin-right: 2rem;
-    }
-  }
-  &__content {
-    display: flex;
-    flex-direction: column;
-    & > * {
-      margin-top: 2rem;
-    }
-    input {
-      width: 30rem;
-    }
-  }
-
-  textarea {
-    width: 74rem;
-    border: 1px solid #dfdfdf;
-    border-radius: 4px;
-  }
-}
