--- src/components/ArticleItem/styles.scss
@@ -1,61 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.article-detail {
-  width: 100%;
-  max-width: 115.2rem;
-  margin: 0 auto;
-  display: grid;
-  gap: 2rem;
-  grid-template-columns: 60% 1fr;
-  grid-auto-rows: minmax(min-content, max-content);
-  padding-top: 2rem;
-
-  &__header {
-    h1 {
-      font-weight: 700;
-    }
-  }
-  &__detail {
-    grid-column: 1 / 2;
-    display: flex;
-    flex-direction: column;
-    & > * {
-      margin-top: 2rem;
-    }
-    img {
-      width: 100%;
-    }
-  }
-  &__informations {
-    display: flex;
-    font-size: 1.6rem;
-    margin-top: 2rem;
-    p:not(:last-of-type) {
-      margin-right: 2rem;
-    }
-  }
-
-  &__text {
-    font-size: 1.6rem;
-  }
-
-  &__sidebar {
-    h4 {
-      margin-top: 3.5rem;
-    }
-  }
-
-  &__item {
-    margin-top: 2.5rem;
-    p {
-      margin-top: 1rem;
-    }
-    a {
-      color: $text-primary;
-      text-decoration: none;
-      &:hover {
-        color: $text-secondary;
-      }
-    }
-  }
-}
