+++ src/components/AvatarItem/images/default_avatar.png
