--- src/components/DisplayArticlesList/styles.scss
@@ -1,16 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.articles {
-  width: 100%;
-  max-width: 115.2rem;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-
-  &__list {
-    width: 100%;
-    margin-top: 4rem;
-    border-top: $border-btn-primary;
-  }
-}
