--- src/components/FormButton/styles.scss
@@ -1,32 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.button {
-  cursor: pointer;
-  width: 17rem;
-  height: 3.5rem;
-  border-radius: $border-r-4px;
-  border: $border-btn-primary;
-  padding: 0.6rem 1.2rem;
-  color: $text-white;
-  font-size: 1.6rem;
-
-  &--primary {
-    background-color: $btn-primary;
-  }
-  &--secondary {
-    background-color: $btn-secondary;
-    border: 1px solid $btn-secondary;
-  }
-
-  &--start {
-    align-self: flex-start;
-  }
-
-  &--end {
-    align-self: flex-end;
-  }
-
-  &:hover {
-    filter: brightness(1.2);
-  }
-}
