+++ src/components/FormComponents/FormLabel/styles.scss
@@ -1,0 +1,6 @@
+@import "../../../styles/abstracts/variables";
+
+.form-label {
+  font-size: 1.6rem;
+  color: $text-primary;
+}
