--- src/components/FormLabel/styles.scss
@@ -1,6 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.form-label {
-  font-size: 1.6rem;
-  color: $text-primary;
-}
