+++ src/components/Header/HeaderLogo/styles.scss
