--- src/components/HeaderAccount/styles.scss
@@ -1,31 +1,0 @@
-@import "../../styles/abstracts/variables";
-
-.header__account {
-  &-menu {
-    display: flex;
-    list-style: none;
-  }
-  &-link {
-    margin-right: 2rem;
-    font-size: 1.6rem;
-    &:link,
-    &:visited {
-      text-decoration: none;
-      color: $btn-primary;
-    }
-    &:hover,
-    &:active {
-      color: $text-primary;
-    }
-  }
-
-  &-avatar {
-    width: 4rem;
-    height: 4rem;
-    border-radius: 50%;
-    box-shadow: $modal-shadow;
-    img {
-      height: 100%;
-    }
-  }
-}
